Force gameState to START at beginning of the game
This is required to eliminate a state bug that would occur
when a new game is started after returning to the main menu
after a game has ended. This was caused because the START state
is only set in the contructor.

This change forces the state to be START at the beginning of the game.